# Conflicts:
#	common/src/main/java/jp/co/soramitsu/common/utils/FlowExt.kt
#	core-db/src/main/java/jp/co/soramitsu/core_db/AppDatabase.kt
#	core-db/src/main/java/jp/co/soramitsu/core_db/di/DbApi.kt
#	core-db/src/main/java/jp/co/soramitsu/core_db/di/DbModule.kt
#	core-db/src/main/java/jp/co/soramitsu/core_db/migrations/Migrations.kt

+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/domain/model/TotalReward.kt
@@ -1,5 +1,5 @@
 package jp.co.soramitsu.feature_staking_impl.domain.model
 
-import java.math.BigDecimal
+import java.math.BigInteger
 
-typealias TotalReward = BigDecimal
+typealias TotalReward = BigInteger
